move external envi file to pytest-simcore
