+++ packages/pytest-simcore/src/pytest_simcore/__init__.py
@@ -1,5 +1,6 @@
 # Collection of tests fixtures for integration testing
 from importlib.metadata import version
+from pathlib import Path
 
 import pytest
 
@@ -7,12 +8,22 @@ __version__: str = version("pytest-simcore")
 
 
 def pytest_addoption(parser):
-    group = parser.getgroup("simcore")
-    group.addoption(
+    simcore_group = parser.getgroup(
+        "simcore", description="options related to pytest simcore"
+    )
+    simcore_group.addoption(
         "--keep-docker-up",
         action="store_true",
         default=False,
         help="Keep stack/registry up after fixtures closes",
+    )
+
+    simcore_group.addoption(
+        "--external-envfile",
+        action="store",
+        type=Path,
+        default=None,
+        help="Path to an env file. Consider passing a link to repo configs, i.e. `ln -s /path/to/osparc-ops-config/repo.config`",
     )
 
     # DUMMY
